+++ application/models/Tracker/Tracker_Favourites_Model.php
@@ -42,6 +42,42 @@ class Tracker_Favourites_Model extends Tracker_Base_Model {
 
 		return $favourites;
 	}
+	public function getAll() : array {
+		$query = $this->db
+			->select('SQL_CALC_FOUND_ROWS
+			          tt.title, tt.title_url,
+			          ts.site, ts.site_class,
+			          tf.chapter, tf.updated_at', FALSE)
+			->from('tracker_favourites AS tf')
+			->join('tracker_chapters AS tc', 'tf.chapter_id = tc.id', 'left')
+			->join('tracker_titles AS tt',   'tc.title_id = tt.id',   'left')
+			->join('tracker_sites AS ts',    'tt.site_id = ts.id',    'left')
+			->where('tc.user_id', $this->User->id) //CHECK: Is this inefficient? Would it be better to have a user_id column in tracker_favourites?
+			->order_by('tf.id DESC')
+			->get();
+
+		$favourites = [];
+		if($query->num_rows() > 0) {
+			foreach($query->result() as $row) {
+				$arrRow = [];
+
+				$arrRow['updated_at'] = $row->updated_at;
+				$arrRow['title']      = $row->title;
+				$arrRow['title_url']  = $this->Tracker->sites->{$row->site_class}->getFullTitleURL($row->title_url);
+
+				$arrRow['site']        = $row->site;
+				$arrRow['chapter']     = $row->chapter;
+
+				$chapterData = $this->Tracker->sites->{$row->site_class}->getChapterData($row->title_url, $row->chapter);
+				$arrRow['chapter_number'] = $chapterData['number'];
+				$arrRow['chapter_url'] = $chapterData['url'];
+
+				$favourites[] = $arrRow;
+			}
+		}
+
+		return $favourites;
+	}
 
 	public function set(string $site, string $title, string $chapter, ?int $userID = NULL) : array {
 		$success = array(