+++ application/models/Tracker/Tracker_Portation_Model.php
@@ -75,4 +75,51 @@ class Tracker_Portation_Model extends Tracker_Base_Model {
 			return $arr;
 		}
 	}
+
+	/**
+	 * Formats a line (passed as a fields  array) as CSV and returns the CSV as a string.
+	 * Adapted from http://us3.php.net/manual/en/function.fputcsv.php#87120
+	 * SEE: http://stackoverflow.com/a/3933816/1168377
+	 *
+	 * @param array  $fields
+	 * @param string $delimiter
+	 * @param string $enclosure
+	 * @param bool   $encloseAll
+	 * @param bool   $nullToMysqlNull
+	 *
+	 * @return string
+	 */
+	public function arrayToCSV(array &$fields, $delimiter = ',', $enclosure = '"', $encloseAll = FALSE, $nullToMysqlNull = FALSE) : string {
+		$delimiter_esc = preg_quote($delimiter, '/');
+		$enclosure_esc = preg_quote($enclosure, '/');
+
+		$output = array();
+		foreach ($fields as $field) {
+			if ($field === NULL && $nullToMysqlNull) {
+				$output[] = 'NULL';
+				continue;
+			}
+
+			// Enclose fields containing $delimiter, $enclosure or whitespace
+			if ($encloseAll || preg_match("/(?:${delimiter_esc}|${enclosure_esc}|\s)/", $field)) {
+				$output[] = $enclosure . str_replace($enclosure, $enclosure . $enclosure, $field) . $enclosure;
+			} else {
+				$output[] = $field;
+			}
+		}
+
+		return implode($delimiter, $output);
+	}
+	public function arrayToCSVRecursive(array &$fields, string $headers = '', $delimiter = ',', $enclosure = '"', $encloseAll = FALSE, $nullToMysqlNull = FALSE) : string {
+		$csvArr = [];
+		if(!empty($headers)) {
+			$csvArr[] = $headers;
+		}
+
+		foreach ($fields as $field) {
+			$csvArr[] = $this->arrayToCSV($field, $delimiter, $enclosure,$encloseAll,$nullToMysqlNull);
+		}
+
+		return implode(PHP_EOL,$csvArr);
+	}
 }